Simplify management of options and arguments
Arguments and options are created in the commands and are subsequently used to get values instead of filtering through available arguments and options recursively up the command tree.

This approach is simpler and matches examples.

Additionally, validation now no longer uses `GetValueOrDefault<T>()` because it invokes validators. Instead, we perform validation similar to `AcceptOnlyFromAmong(...)`